# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ARCode Clone API
  description: API REST complète pour l'application ARCode Clone iOS
  version: 1.0.0
  contact:
    name: ARCode Support
    email: support@ar-code.com
    url: https://ar-code.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.ar-code.com
    description: Production
  - url: https://staging-api.ar-code.com
    description: Staging
  - url: http://localhost:8080
    description: Local Development
...
paths:
...
  /api/v1/ai/generation/txt2img:
    post:
      tags:
        - AI
      summary: Generate Image from Text
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Txt2ImgRequest'
      responses:
        '200':
          description: Image generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIGenerationResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Txt2ImgRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
        negative_prompt:
          type: string
        steps:
          type: integer
          default: 20
        cfg_scale:
          type: number
          default: 7.5
        width:
          type: integer
          default: 512
        height:
          type: integer
          default: 512
        model:
          type: string
          default: sd-v1-5
    AIGenerationResponse:
      type: object
      properties:
        image_id:
          type: string
          format: uuid
        image_url:
          type: string
          format: uri
        seed:
          type: integer
        processing_time:
          type: number
...